processing order and guest user cart

+++ store/models.py
@@ -17,7 +17,7 @@ class Customer(models.Model):
 class Product(models.Model):
     name = models.CharField(max_length=254)
     description = models.CharField(max_length=254, null=True, blank=True)
-    price = models.FloatField()
+    price = models.DecimalField(decimal_places=2, max_digits=7)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
